requirements.txt
# AI/ML Core Packages
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# LangChain Packages
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.8
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# LLM API Packages
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Streamlit Frontend Packages
streamlit>=1.28.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-echarts>=0.4.0

# Other Dependencies
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0